pipeline/requirements.txt
google-cloud-bigquery==3.27.0
google-auth==2.38.0
openai==1.73.0
hdbscan==0.8.41
scikit-learn==1.6.1
numpy==2.2.4
umap-learn==0.5.7
tenacity==9.1.2
